Untested client example and restructure go packages in root
